+++ modules/No.StoreButler.LedgerManagement/src/No.StoreButler.LedgerManagement.Application.Contracts/LedgerManagement/Dtos/UpdatePayTradeInput.cs
@@ -18,7 +18,7 @@ namespace No.StoreButler.LedgerManagement.LedgerManagement.Dtos
         /// <summary>
         /// 交易金额
         /// </summary>
-        public long PayAmount { get; set; }
+        public decimal PayAmount { get; set; }
         
         /// <summary>
         /// 资金流向
